+++ src/Option.ts
@@ -42,7 +42,7 @@ import type { Semigroup } from './Semigroup'
 import * as separated from './Separated'
 import type { Show } from './Show'
 import * as traversable from './Traversable'
-import * as witherable from './Witherable'
+import * as filterableE from './FilterableE'
 import * as zero_ from './Zero'
 import type { Separated } from './Separated'
 import type { Eq } from './Eq'
@@ -916,38 +916,36 @@ export const sequence: <F extends HKT>(
   /*#__PURE__*/ traversable.sequence<OptionF>(Traversable)
 
 /**
- * @category Witherable
+ * @category FilterableE
  * @since 3.0.0
  */
-export const wither: <F extends HKT>(
+export const filterMapE: <F extends HKT>(
   F: applicative.Applicative<F>
 ) => <A, S, R, W, E, B>(
   f: (a: A) => Kind<F, S, R, W, E, Option<B>>
-) => (ta: Option<A>) => Kind<F, S, R, W, E, Option<B>> = /*#__PURE__*/ witherable.witherDefault(
+) => (ta: Option<A>) => Kind<F, S, R, W, E, Option<B>> = /*#__PURE__*/ filterableE.filterMapEDefault(
   Traversable,
   Compactable
 )
 
 /**
- * @category Witherable
+ * @category FilterableE
  * @since 3.0.0
  */
-export const wilt: <F extends HKT>(
+export const partitionMapE: <F extends HKT>(
   F: applicative.Applicative<F>
 ) => <A, S, R, W, E, B, C>(
   f: (a: A) => Kind<F, S, R, W, E, Either<B, C>>
-) => (wa: Option<A>) => Kind<F, S, R, W, E, Separated<Option<B>, Option<C>>> = /*#__PURE__*/ witherable.wiltDefault(
-  Traversable,
-  Compactable
-)
+) => (wa: Option<A>) => Kind<F, S, R, W, E, Separated<Option<B>, Option<C>>> =
+  /*#__PURE__*/ filterableE.partitionMapEDefault(Traversable, Compactable)
 
 /**
  * @category instances
  * @since 3.0.0
  */
-export const Witherable: witherable.Witherable<OptionF> = {
-  wither,
-  wilt
+export const FilterableE: filterableE.FilterableE<OptionF> = {
+  filterMapE,
+  partitionMapE
 }
 
 /**